refactor(H): 按maintainer意见将release_orgasm_edge_now迁至second_behavior并合并群交退出遍历
- 工具函数release_orgasm_edge_now从Script/Settle/default.py迁至Script/Design/second_behavior.py（高潮相关工具函数所在处），原三处调用改为second_behavior.release_orgasm_edge_now引用
- handle_group_sex_end_h_add_hpmp_max中新增的角色遍历与原遍历完全一致，合并进原遍历、置于绝顶输出统计之前，去掉多余的额外遍历

纯重构，不改结算语义
